refactor(core-ui-theme): change to kmp
