+++ app-k9mail/src/main/kotlin/app/k9mail/provider/K9FeatureThemeProvider.kt
@@ -2,7 +2,7 @@ package app.k9mail.provider
 
 import androidx.compose.runtime.Composable
 import app.k9mail.core.ui.compose.theme2.k9mail.K9MailTheme2
-import app.k9mail.core.ui.theme.api.FeatureThemeProvider
+import net.thunderbird.core.ui.theme.api.FeatureThemeProvider
 
 internal class K9FeatureThemeProvider : FeatureThemeProvider {
     @Composable
