+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/provider/ProviderModule.kt
@@ -1,10 +1,10 @@
 package net.thunderbird.android.provider
 
-import app.k9mail.core.ui.theme.api.FeatureThemeProvider
-import app.k9mail.core.ui.theme.api.ThemeProvider
 import com.fsck.k9.preferences.FilePrefixProvider
 import net.thunderbird.core.common.provider.AppNameProvider
 import net.thunderbird.core.common.provider.BrandNameProvider
+import net.thunderbird.core.ui.theme.api.FeatureThemeProvider
+import net.thunderbird.core.ui.theme.api.ThemeProvider
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.binds
 import org.koin.dsl.module
