+++ app-thunderbird/src/main/kotlin/net/thunderbird/android/provider/TbFeatureThemeProvider.kt
@@ -2,7 +2,7 @@ package net.thunderbird.android.provider
 
 import androidx.compose.runtime.Composable
 import app.k9mail.core.ui.compose.theme2.thunderbird.ThunderbirdTheme2
-import app.k9mail.core.ui.theme.api.FeatureThemeProvider
+import net.thunderbird.core.ui.theme.api.FeatureThemeProvider
 
 internal class TbFeatureThemeProvider : FeatureThemeProvider {
     @Composable
