+++ core/ui/theme/api/src/commonMain/kotlin/net/thunderbird/core/ui/theme/api/Theme.kt
@@ -1,4 +1,4 @@
-package app.k9mail.core.ui.theme.api
+package net.thunderbird.core.ui.theme.api
 
 enum class Theme {
     LIGHT,
