+++ core/ui/theme/manager/src/main/java/net/thunderbird/core/ui/theme/manager/ThemeManager.kt
@@ -4,9 +4,6 @@ import android.content.Context
 import android.content.res.Configuration
 import androidx.annotation.StyleRes
 import androidx.appcompat.app.AppCompatDelegate
-import app.k9mail.core.ui.theme.api.Theme
-import app.k9mail.core.ui.theme.api.ThemeManager
-import app.k9mail.core.ui.theme.api.ThemeProvider
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.distinctUntilChanged
@@ -18,6 +15,9 @@ import net.thunderbird.core.preferences.AppTheme
 import net.thunderbird.core.preferences.GeneralSettings
 import net.thunderbird.core.preferences.GeneralSettingsManager
 import net.thunderbird.core.preferences.SubTheme
+import net.thunderbird.core.ui.theme.api.Theme
+import net.thunderbird.core.ui.theme.api.ThemeManager
+import net.thunderbird.core.ui.theme.api.ThemeProvider
 
 class ThemeManager(
     private val context: Context,
