+++ feature/launcher/src/main/kotlin/app/k9mail/feature/launcher/ui/FeatureLauncherApp.kt
@@ -8,8 +8,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.navigation.compose.rememberNavController
 import app.k9mail.core.ui.compose.designsystem.atom.Surface
-import app.k9mail.core.ui.theme.api.FeatureThemeProvider
 import app.k9mail.feature.launcher.navigation.FeatureLauncherNavHost
+import net.thunderbird.core.ui.theme.api.FeatureThemeProvider
 import org.koin.compose.koinInject
 
 @Composable
