+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/helper/HtmlSettingsProvider.kt
@@ -1,8 +1,8 @@
 package com.fsck.k9.ui.helper
 
-import app.k9mail.core.ui.theme.api.Theme
 import com.fsck.k9.K9
 import com.fsck.k9.message.html.HtmlSettings
+import net.thunderbird.core.ui.theme.api.Theme
 import net.thunderbird.core.ui.theme.manager.ThemeManager
 
 class HtmlSettingsProvider(private val themeManager: ThemeManager) {
