+++ legacy/ui/legacy/src/main/java/com/fsck/k9/view/WebViewConfigProvider.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.view
 
-import app.k9mail.core.ui.theme.api.Theme
 import com.fsck.k9.K9
+import net.thunderbird.core.ui.theme.api.Theme
 import net.thunderbird.core.ui.theme.manager.ThemeManager
 
 class WebViewConfigProvider(private val themeManager: ThemeManager) {
